Rename reset to seed and seed products from CSV
Rename products/reset.py to products/seed.py and update API to seed the product table from an in-package CSV. Change endpoint from POST /products/reset to GET /products/seed, move CSV to app/api/products/start_data.csv and resolve its absolute path. Map CSV 'desc' field to DB 'title' on insert, update SQL to use title, and guard against empty query results (handle cur.description None). Update imports and router includes to use products_router/seed_router names. Remove legacy product-list root logic and clean up tests by deleting obsolete echo tests.
