+++ app/api/products/seed.py
@@ -1,14 +1,16 @@
-import os
+from app import __version__
+import os, time
 import csv
 from fastapi import APIRouter, status
 from app.api.db import get_db_connection
 
 router = APIRouter()
 
-CSV_PATH = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(__file__))), 'tests', 'csv', 'small.csv')
+CSV_PATH = os.path.join(os.path.dirname(__file__), 'start_data.csv')
+CSV_PATH = os.path.abspath(CSV_PATH)
 
-@router.post("/products/reset", status_code=status.HTTP_200_OK)
-def reset_products() -> dict:
+@router.get("/products/seed", status_code=status.HTTP_200_OK)
+def seed_products() -> dict:
     """Delete and recreate the product table, then seed with CSV data."""
     conn_gen = get_db_connection()
     conn = next(conn_gen)
@@ -20,7 +22,7 @@ def reset_products() -> dict:
             id SERIAL PRIMARY KEY,
             Params TEXT,
             item INTEGER,
-            desc TEXT,
+            title TEXT,
             UOS TEXT,
             Pack_Description TEXT,
             Hierarchy1 TEXT,
@@ -45,13 +47,15 @@ def reset_products() -> dict:
         reader = csv.DictReader(csvfile)
         rows = [row for row in reader]
         for row in rows:
+            # Map 'desc' from CSV to 'title' for DB
+            row['title'] = row.pop('desc')
             cur.execute(
                 """
                 INSERT INTO product (
-                    Params, item, desc, UOS, Pack_Description, Hierarchy1, Hierarchy2, Hierarchy3, UOP,
+                    Params, item, title, UOS, Pack_Description, Hierarchy1, Hierarchy2, Hierarchy3, UOP,
                     sSell1, sSell2, sSell3, sSell4, sSell5, pack1, pack2, pack3, pack4, pack5, EAN
                 ) VALUES (
-                    %(Params)s, %(item)s, %(desc)s, %(UOS)s, %(Pack_Description)s, %(Hierarchy1)s, %(Hierarchy2)s, %(Hierarchy3)s, %(UOP)s,
+                    %(Params)s, %(item)s, %(title)s, %(UOS)s, %(Pack_Description)s, %(Hierarchy1)s, %(Hierarchy2)s, %(Hierarchy3)s, %(UOP)s,
                     %(sSell1)s, %(sSell2)s, %(sSell3)s, %(sSell4)s, %(sSell5)s, %(pack1)s, %(pack2)s, %(pack3)s, %(pack4)s, %(pack5)s, %(EAN)s
                 )
                 """,
@@ -59,51 +63,19 @@ def reset_products() -> dict:
             )
     conn.commit()
     cur.execute('SELECT * FROM product;')
-    products = [dict(zip([desc[0] for desc in cur.description], row)) for row in cur.fetchall()]
-    cur.close()
-    conn.close()
-    return {"message": "Product table reset and seeded from CSV.", "data": products}
-import os, time
-import psycopg2
-from dotenv import load_dotenv
-from app import __version__
-
-router = APIRouter()
-
-@router.get("/products")
-def root() -> dict:
-    """Return a structured welcome message for the API root, including product data."""
-    load_dotenv()
-    conn = psycopg2.connect(
-        host=os.getenv('DB_HOST'),
-        port=os.getenv('DB_PORT', '5432'),
-        dbname=os.getenv('DB_NAME'),
-        user=os.getenv('DB_USER'),
-        password=os.getenv('DB_PASSWORD')
-    )
-    cur = conn.cursor()
-    cur.execute('SELECT id, name, description, price, in_stock, created_at FROM product;')
-    products = [
-        {
-            "id": row[0],
-            "name": row[1],
-            "description": row[2],
-            "price": float(row[3]),
-            "in_stock": row[4],
-            "created_at": row[5].isoformat() if row[5] else None
-        }
-        for row in cur.fetchall()
-    ]
+    if cur.description is None:
+        products = []
+    else:
+        columns = [desc[0] for desc in cur.description]
+        products = [dict(zip(columns, row)) for row in cur.fetchall()]
     cur.close()
     conn.close()
 
-    load_dotenv()
     base_url = os.getenv("BASE_URL", "http://localhost:8000")
-
     epoch = int(time.time() * 1000)
     meta = {
         "severity": "success",
-        "title": "Product List",
+        "title": "Product table seeded",
         "version": __version__,
         "base_url": base_url,
         "time": epoch,
