+++ app/api/routes.py
@@ -13,9 +13,9 @@ router = APIRouter()
 from app.api.root import router as root_router
 from app.api.health import router as health_router
 from app.api.products.products import router as products_router
-from app.api.products.reset import router as reset_router
+from app.api.products.seed import router as seed_router
 
 router.include_router(root_router)
 router.include_router(health_router)
 router.include_router(products_router)
-router.include_router(reset_router)
+router.include_router(seed_router)
