+++ tests/test_routes.py
@@ -25,22 +25,3 @@ def test_health_returns_ok() -> None:
     assert response.status_code == 200
     assert response.json() == {"status": "ok"}
 
-
-def test_echo_returns_message() -> None:
-    """POST /echo should return the provided message."""
-    response = client.post("/echo", json={"message": "hello"})
-    assert response.status_code == 200
-    assert response.json() == {"echo": "hello"}
-
-
-def test_echo_empty_string() -> None:
-    """POST /echo with an empty string should echo back an empty string."""
-    response = client.post("/echo", json={"message": ""})
-    assert response.status_code == 200
-    assert response.json() == {"echo": ""}
-
-
-def test_echo_missing_body_returns_422() -> None:
-    """POST /echo without a body should return 422 Unprocessable Entity."""
-    response = client.post("/echo", json={})
-    assert response.status_code == 422
